% Generated by build script
\name{cr_funders}
\alias{cr_funders}
\title{Search Crossref funders}
\usage{
cr_funders()
}
\arguments{
\item{query}{Search query (funder name)}
\item{rows}{Number of results (default 10)}
}
\value{tibble: id, name, location, alt_names, uri, work_count}
\description{Search Crossref funders}
