% Generated by build script
\name{cr_journals}
\alias{cr_journals}
\title{Search Crossref journals}
\usage{
cr_journals()
}
\arguments{
\item{query}{Search query string (journal title)}
\item{rows}{Number of results. Default 10.}
}
\value{tibble: issn, title, publisher, subjects, total_dois}
\description{Search Crossref journals}
